circleBoxesRepository ended. New tests for circleBoxesRepository. Refactoring on int types

+++ src/main/scala/com/fl3sc0b/whitechapelai/board/CircleBox.scala
@@ -1,5 +1,5 @@
 package com.fl3sc0b.whitechapelai.board
 
-case class CircleBox(id: String, number: Byte, red: Boolean, adjacentSquares: List[String]) extends Box{
+case class CircleBox(id: String, number: Short, red: Boolean, adjacentSquares: List[String]) extends Box{
   def constructId: String = id
 }
